+++ app/src/main/java/com/stonymoon/bboard/songdashboard/SongDashboardContract.java
@@ -2,14 +2,13 @@ package com.stonymoon.bboard.songdashboard;
 
 import com.stonymoon.bboard.base.BasePresenter;
 import com.stonymoon.bboard.base.BaseView;
-import com.stonymoon.bboard.bean.ItunesBean;
-import com.stonymoon.bboard.bean.SongBean;
+
 
 import java.util.List;
 
 public interface SongDashboardContract {
     interface View extends BaseView<Presenter> {
-        void showChart(List<List<Long>> list);
+        void showChart(List<List<Double>> list);
 
         void showError();
